pyproject.toml
[project]
name = "discord-video-embed-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
    "yt-dlp>=2025.6.9",
]
